The web page control of the robot
This is the web app wich provides commands button to the user to take control of the robot.
Each action that the robot is able to make has a submit button wich send a request to the server by a get method(although not secured, but will be replace by a post method after). Each button is action named(for eg the button to to MOVE FORWARD is labeled "FORWAD" so in french "AVANCER", to MOVE BACKWARD is in french "RECULER"), and is filled by a picture wich tells about the action.
When a button is clicked, a get request formed by the two parameters "led" and "led_value" are sent to the python server runing on the MR3020 router. For now ,only the "led_value" parameter is used. Its values represent a precise action transfered to the MSP430 microcontroler vi serial link, so the microcontroler knows what relay to activate for the action requested.
Note that this web page is styled by a css file named app.css available in this repository.